refactor(core): Reorganize core module structure for better maintainability
- Split monolithic `models.py` into `models/validation.py` and `models/metrics.py`
- Move protocols to `base/` directory (`analyzer.py`, `protocols.py`)
- Consolidate constraints under `logic/comparators.py`
- Relocate profiler and schema services to `services/` subdirectory
- Update all imports across engines to reflect new structure
- Remove empty/unused `analyzers/` and `sink/` directories
- Fix PySpark engine initialization of `_dq_errors` with proper schema

This reorganization establishes a cleaner separation between core models, base protocols, business logic, and services, making the codebase more maintainable and scalable.
