+++ sumeh/core/constraints/__init__.py
@@ -15,7 +15,7 @@ Usage:
     >>> print(result.status)  # PASS or FAIL
 """
 
-from sumeh.core.constraints.comparators import (
+from sumeh.core.logic.comparators import (
     CompletenessConstraint,
     MultiFieldCompletenessConstraint,
 )
