+++ sumeh/core/services/profiler/profiler.py
@@ -10,7 +10,7 @@ import time
 from typing import Any, Dict, Optional
 
 from sumeh.core.rules.rule_model import RuleDef
-from sumeh.core.schema.validator import extract_schema
+from sumeh.core.services.schema.validator import extract_schema
 
 
 class DataProfiler:
