+++ sumeh/core/services/schema/__init__.py
@@ -3,7 +3,7 @@ Sumeh Schema Gatekeeper.
 Validates DataFrame schemas before data quality validation.
 """
 
-from sumeh.core.schema.validator import validate, extract_schema
-from sumeh.core.schema.models import SchemaDef, ColumnDef, SchemaReport
+from sumeh.core.services.schema.validator import validate, extract_schema
+from sumeh.core.services.schema.models import SchemaDef, ColumnDef, SchemaReport
 
 __all__ = ["validate", "extract_schema", "SchemaDef", "ColumnDef", "SchemaReport"]
