+++ sumeh/engines/bigquery/engine.py
@@ -11,7 +11,7 @@ try:
 except ImportError:
     bigquery = None
 
-from sumeh.core.models import ValidationReport
+from sumeh.core.models.validation import ValidationReport
 from sumeh.core.rules.rule_model import RuleDef
 from sumeh.engines.sql_core.compiler import compile_rules_to_sql
 from sumeh.engines.sql_core.validator import validate_results
