+++ sumeh/engines/duckdb/engine.py
@@ -8,7 +8,7 @@ from typing import List, Union, Any, Optional
 
 import duckdb
 
-from sumeh.core.models import ValidationReport
+from sumeh.core.models.validation import ValidationReport
 from sumeh.core.rules.rule_model import RuleDef
 from sumeh.engines.sql_core.compiler import compile_rules_to_sql
 from sumeh.engines.sql_core.validator import validate_results
