+++ sumeh/engines/pandas/analyzers.py
@@ -7,7 +7,7 @@ All row-level + table-level analyzers for pandas engine.
 import numpy as np
 import pandas as pd
 
-from sumeh.core.models import MetricResult
+from sumeh.core.models.metrics import MetricResult
 from sumeh.core.rules.rule_model import RuleDef
 
 # ============================================================================
